+++ src/Suru.Compiler/Parse/Ast/SourcePosition.cs
@@ -1,0 +1,6 @@
+namespace Suru.Compiler.Parse.Ast;
+
+public readonly record struct SourcePosition(int Line, int Column)
+{
+    public override string ToString() => $"({Line},{Column})";
+}